fix: extract AiCategorizationError to shared module for instanceof compat
Move AiCategorizationError into its own lightweight module so both
the real ai-categorizer and mock can import the same class. This
preserves instanceof checks in service.ts while avoiding the heavy
dynamic import that caused test timeouts.
